backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
torch
numpy
python-multipart
pdfplumber
python-docx